# swagger.yaml
openapi: 3.0.0
info:
  title: invaPlay Backend API
  version: 1.0.0
  description: API documentation for the invaPlay Backend application, managing users, students, schools, classes, and grades.
servers:
  - url: http://localhost:5000 # Assuming the API is served from the root
  - url: https://api.innovplay.online/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /classes:
    get:
      summary: Get list of classes with school and grade details
      responses:
        '200':
          description: A list of classes with related school and grade details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/schemas/ClassDetails'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '500':
          $ref: '#/responses/InternalServerError'
    post:
      summary: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                school_id:
                  type: string
                grade_id:
                  type: string
                  nullable: true
              required:
                - name
                - school_id
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  class:
                    $ref: '#/schemas/Class'
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '500':
          $ref: '#/responses/InternalServerError'
...
  /school:
    get:
      summary: Get list of schools with city, student and class counts
      responses:
        '200':
          description: A list of schools with related city information and student/class counts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    domain:
                      type: string
                    address:
                      type: string
                    city_id:
                      type: string
                    created_at:
                      type: string
                      format: date-time
                      nullable: true
                    students_count:
                      type: integer
                    classes_count:
                      type: integer
                    city:
                      $ref: '#/schemas/CityDetails'
                      nullable: true
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/InternalServerError'
    post:
      summary: Create a new school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                domain:
                  type: string
                  nullable: true
                address:
                  type: string
                  nullable: true
                city_id:
                  type: string
              required:
                - name
                - city_id
      responses:
        '201':
          description: School created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  id:
                    type: string
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '500':
          $ref: '#/responses/InternalServerError'
...
responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem explicando o motivo da falha de autenticação.
                example: Token inválido ou expirado.
    ForbiddenError:
      description: User does not have the necessary permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Mensagem indicando que o usuário não tem permissão.
                example: Acesso negado para o recurso solicitado.
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem sobre o recurso não encontrado.
                example: Recurso solicitado não foi encontrado.
    BadRequestError:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem descrevendo o erro de validação.
                example: Dados inválidos enviados na requisição.
              details:
                type: object # Can be a string or a more detailed object
                description: Detalhes adicionais sobre o erro.
                example:
                  campo: email
                  problema: Formato inválido.
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Mensagem genérica de erro.
                example: Ocorreu um erro inesperado.
              details:
                type: string
                description: Descrição detalhada do erro para depuração.
                example: Consulte os logs do servidor para mais informações.